Add docs, restructure for dartdoc tool
