--- lib/actions/set_analytics_collection_enabled.dart
@@ -1,7 +1,0 @@
-import 'package:analyticsx/analytics_action.dart';
-
-class SetAnalyticsCollectionEnabled extends AnalyticsAction {
-  final bool enabled;
-
-  SetAnalyticsCollectionEnabled({required this.enabled});
-}
