--- lib/actions/set_screen.dart
@@ -1,8 +1,0 @@
-import 'package:analyticsx/analytics_action.dart';
-
-class SetScreen extends AnalyticsAction {
-  final String screenName;
-  final String screenClassOverride;
-
-  SetScreen(this.screenName, this.screenClassOverride);
-}
