--- lib/analytics_action.dart
@@ -1,1 +1,0 @@
-abstract class AnalyticsAction {}
