+++ lib/src/actions/set_analytics_collection_enabled.dart
@@ -1,0 +1,10 @@
+import 'package:analyticsx/src/analytics_action.dart';
+
+/// Represents an action to disable/enable analytics collection
+///
+/// This might be useful for a screen recording analytics provider that a developer doesn't want PII leaked to
+class SetAnalyticsCollectionEnabled extends AnalyticsAction {
+  final bool enabled;
+
+  SetAnalyticsCollectionEnabled({required this.enabled});
+}
