+++ lib/src/analytics_action.dart
@@ -1,0 +1,5 @@
+/// Abstract class to represent an action that an analytics vendor might be interested in.
+///
+/// Concrete implementations will hold all information required for the vendor to act on the action. See [TrackEvent]
+/// for an example implementation.
+abstract class AnalyticsAction {}
